+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [2.0.0] - 2017-08-13
 - Changed `set` method: last argument now is an object, see documentation.
-- Added expire date
+- Added expiry date
 
 ## [1.0.1] - 2017-08-12
 - First release