+++ CHANGELOG.md
@@ -5,8 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+- Library loading was changed:
+    - **Now you need to instantiate the `InCache` object**
 - Added cache writing to disk
 - Added `setConfig` method
+- Added `addTo` method
+- Added `prependTo` method
+- Added `updateIn` method
+- Added `removeFrom` method
 
 ## [2.0.0] - 2017-08-13
 - Changed `set` method: last argument now is an object, see documentation.