+++ CHANGELOG.md
@@ -4,15 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [unreleased]
+## [3.0.0]
 - Library loading was changed:
-    - **Now you need to instantiate the `InCache` object**
+    - **Now you need to instantiate the `new InCache` object**
 - Added cache writing to disk
 - Added `setConfig` method
 - Added `addTo` method
 - Added `prependTo` method
 - Added `updateIn` method
 - Added `removeFrom` method
+- Added global record configuration
 
 ## [2.0.0] - 2017-08-13
 - Changed `set` method: last argument now is an object, see documentation.